[Feat] 고객 완료, 사진 하는 중

+++ src/components/CastComponent.js
@@ -8,6 +8,7 @@ export default function CastComponent(props) {
   const setCastList = props.setCastList;
 
   const [castModifyOpen, setCastModifyOpen] = useState(false);
+  const [isLoading, setLoading] = useState(false);
 
   const showCastModify = () => {
     setCastModifyOpen(true);
@@ -26,15 +27,24 @@ export default function CastComponent(props) {
       },
     };
 
+    const yesOrNo = window.confirm('인물을 삭제하시겠습니까?');
+    if (yesOrNo === false) {
+      return;
+    }
+
     try {
-      console.log('staffToken', token);
+      setLoading(true);
+
       const response = await serverapi.delete(api, options);
       console.log('response', response.data);
 
+      alert('삭제되었습니다');
       setCastList(response.data);
     } catch (error) {
       console.log(error);
       alert(error.response.data.message);
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -62,8 +72,14 @@ export default function CastComponent(props) {
         </Modal>
       </td>
       <td>
-        <button class='btn btn-danger' onClick={() => deleteCast(cast.castId)}>
-          삭제
+        <button class='btn btn-danger' onClick={deleteCast(cast.castId)}>
+          {isLoading ? (
+            <div className='spinner-border' role='status'>
+              <span className='sr-only' />
+            </div>
+          ) : (
+            <span>삭제</span>
+          )}
         </button>
       </td>
     </tr>
